refactor: computer
